backv2-github/DevDocs/FinDocRAG/requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
pymupdf==1.23.8
pdf2image==1.16.3
pytesseract==0.3.10
pandas==2.1.1
numpy==1.26.0
google-generativeai==0.3.1
python-dotenv==1.0.0  # For loading environment variables
requests==2.31.0
pydantic==2.5.2
langchain==0.0.335
langchain-google-genai==0.0.5
pillow==10.1.0
camelot-py==0.11.0
opencv-python-headless==4.8.1.78
tabula-py==2.8.2
openpyxl==3.1.2
pdfplumber==0.10.2
scikit-learn==1.3.2
matplotlib==3.8.0
seaborn==0.13.0
